requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.1
Flask-WTF==1.2.1
gunicorn==20.0.4
h11==0.14.0
icmplib==3.0.4
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
networkx==3.2.1
outcome==1.3.0.post0
packaging==24.0
pluggy==1.4.0
prometheus_client==0.20.0
psutil==5.9.8
PyMySQL==1.1.0
PySocks==1.7.1
pytest==8.1.1
python-dotenv==1.0.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
zipp==3.17.0
